Add files via upload
These are the R functions that the authors of the manuscript for the Environmental Challenge journal have created. For the first time, these functions facilitate curve fitting to the grazing orbits and parameter estimation, which is a major breakthrough presented in the manuscript. We have demonstrated the curve fitting beyond any reasonable doubt, making it one of the manuscript's greatest contributions to the knowledge of herd mobility and mathematical formulation, particularly for the double Gompertz function.