Add files via upload
In an article about Environmental Challenges, Figure 5 illustrates the successful utilization of the double Gompertz function to fit a curve to a grazing path. This highlights the ability of the double Gompertz function to produce indicators that can be easily interpreted from the perspective of pastoralists. The authors were able to fit the double Gompertz function to one of the numerous grazing paths, which is a noteworthy accomplishment.