Add files via upload
The manuscript for Environmental Challenge Journal contains codes for Table 2, which provides monthly abundance analysis and return output for interpretation. The table includes a mix of grazing orbit use and various factors that influence monthly abundance. It highlights the use of orbits and emphasizes the need for a certain level of confidence to use grazing orbits to identify grazing places.